+++ package.json
@@ -12,7 +12,7 @@
   "dependencies": {
     "@tailwindcss/vite": "^4.1.17",
     "dayjs": "^1.11.19",
-    "firebase": "^12.5.0",
+    "firebase": "^12.6.0",
     "lucide-react": "^0.552.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
