+++ src/components/icons/DecisionIcon.tsx
@@ -1,0 +1,5 @@
+export const DecisionIcon = () => (
+    <svg width="20" height="20" viewBox="-0.5 -0.5 5 5" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <path d="M2.08333 0.416668C1.82707 0.416504 1.57424 0.475509 1.34452 0.589084C1.11481 0.702659 0.914414 0.867738 0.758958 1.07146L1.25 1.5625H0V0.312501L0.462292 0.774793C0.657362 0.532652 0.904259 0.337347 1.1848 0.203259C1.46535 0.0691706 1.77239 -0.000286837 2.08333 8.90288e-07C3.23396 8.90288e-07 4.16667 0.932709 4.16667 2.08333H3.75C3.75 1.64131 3.57441 1.21738 3.26184 0.904823C2.94928 0.592262 2.52536 0.416668 2.08333 0.416668ZM0.416667 2.08333C0.416658 2.43187 0.525919 2.77165 0.729079 3.05485C0.932239 3.33806 1.21907 3.55045 1.54923 3.66214C1.87939 3.77384 2.23626 3.77921 2.56963 3.67752C2.90301 3.57582 3.19611 3.37217 3.40771 3.09521L2.91667 2.60417H4.16667V3.85417L3.70438 3.39188C3.50931 3.63402 3.26241 3.82932 2.98186 3.96341C2.70132 4.0975 2.39427 4.16696 2.08333 4.16667C0.932708 4.16667 0 3.23396 0 2.08333H0.416667Z" fill="#FA7315"/>
+    </svg>
+);
