+++ src/components/icons/PlaceIcon.tsx
@@ -1,0 +1,6 @@
+export const PlaceIcon = () => (
+    <svg width="23" height="23" viewBox="-0.3 -0.2 5 5" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <path d="M4.04003 1.97304C4.04003 3.25552 2.42401 4.3407 2.42401 4.3407C2.42401 4.3407 0.807983 3.25552 0.807983 1.97304C0.807983 1.55441 0.978243 1.15292 1.28131 0.856908C1.58437 0.560892 1.99541 0.394592 2.42401 0.394592C2.8526 0.394592 3.26365 0.560892 3.56671 0.856908C3.86977 1.15292 4.04003 1.55441 4.04003 1.97304Z" stroke="#31C867" strokeWidth="0.416667"/>
+        <path d="M3.03001 1.97302C3.03001 2.13001 2.96616 2.28056 2.85252 2.39157C2.73887 2.50257 2.58473 2.56494 2.424 2.56494C2.26328 2.56494 2.10914 2.50257 1.99549 2.39157C1.88184 2.28056 1.81799 2.13001 1.81799 1.97302C1.81799 1.81603 1.88184 1.66548 1.99549 1.55447C2.10914 1.44347 2.26328 1.3811 2.424 1.3811C2.58473 1.3811 2.73887 1.44347 2.85252 1.55447C2.96616 1.66548 3.03001 1.81603 3.03001 1.97302Z" stroke="#31C867" strokeWidth="0.416667"/>
+    </svg>
+);
