+++ src/components/icons/TimeIcon.tsx
@@ -1,0 +1,6 @@
+export const TimeIcon = () => (
+    <svg width="25" height="25" viewBox="-0.2 -0.2 6 6" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <path d="M4.56701 2.25528V1.12764H0.652466V2.25528M4.56701 2.25528H0.652466M4.56701 2.25528V4.7361H0.652466V2.25528M1.52236 1.12764V0.45105M3.69711 1.12764V0.45105" stroke="#6059E8" strokeWidth="0.25" strokeLinecap="square"/>
+        <path d="M3.37095 2.98828L2.4482 3.9452L1.98694 3.46663" stroke="#6059E8" strokeWidth="0.25" strokeLinecap="square"/>
+    </svg>
+);
